pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "raagdosa"
version = "10.0.0"
description = "Deterministic music library cleanup for DJs and collectors — CLI-first, safe-by-default, undoable."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
authors = [
  { name = "RaagDosa Contributors" }
]
keywords = ["music", "library", "dj", "mp3", "flac", "metadata", "organise", "organize"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Multimedia :: Sound/Audio",
  "Topic :: Utilities",
]
dependencies = [
  "mutagen>=1.46",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7", "pytest-cov"]

[project.scripts]
raagdosa = "raagdosa:main"

[project.urls]
Homepage = "https://github.com/raagdosa/raagdosa"
Repository = "https://github.com/raagdosa/raagdosa"
"Bug Tracker" = "https://github.com/ajeshh/raagdosa/issues"
Changelog = "https://github.com/ajeshh/raagdosa/blob/main/CHANGELOG.md"

[tool.setuptools]
packages = ["raagdosa"]
py-modules = ["raagdosa_scanner", "raagdosa_tags"]
